Day 8 Task 1 Completed

+++ Entity/Concrete/Car.cs
@@ -13,7 +13,7 @@ namespace Entity.Concrete
         public int BrandId { get; set; }
         public int ColorId { get; set; }
         public DateTime ModelYear { get; set; }
-        public float DailyPrice { get; set; }
+        public decimal DailyPrice { get; set; }
         public string Description { get; set; }
 
     }
